Add removers to IlmSession
Added new traits for instructors, learners and instructor groups and
added them to entities.
