+++ src/Ilios/CoreBundle/Traits/InstructorsEntityInterface.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Ilios\CoreBundle\Traits;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+
+use Ilios\CoreBundle\Entity\UserInterface;
+
+/**
+ * Interface InstructorsEntityInterface
+ * @package Ilios\CoreBundle\Traits
+ */
+interface InstructorsEntityInterface
+{
+    /**
+     * @param Collection $instructors
+     */
+    public function setInstructors(Collection $instructors);
+
+    /**
+     * @param UserInterface $instructor
+     */
+    public function addInstructor(UserInterface $instructor);
+
+    /**
+     * @param UserInterface $instructor
+     */
+    public function removeInstructor(UserInterface $instructor);
+
+    /**
+    * @return UserInterface[]|ArrayCollection
+    */
+    public function getInstructors();
+}
